+++ tests/test_cli.py
@@ -1,13 +1,14 @@
-import pdb
+import sys
 
 import pytest
 from click.testing import CliRunner
 
-from mio.cli.config import config
+from mio.cli.config import config, _list
 from mio.cli.stream import capture
 from mio import Config
 from mio.utils import hash_video
 from mio.models import config as _config_mod
+from .conftest import CONFIG_DIR
 
 from .conftest import DATA_DIR
 
@@ -99,6 +100,42 @@ def test_cli_config_user_path(set_env, set_user_yaml):
     runner = CliRunner()
     result = runner.invoke(config, ["user", "path"])
     assert str(user_config_path) in result.output
+
+
+def test_cli_config_list():
+    """
+    mio config list should list all the configs in the user directory and provided by mio
+    """
+    runner = CliRunner()
+    result = runner.invoke(_list, color=False)
+
+    # not testing for the literal table structure, but we should have headers and some table characters
+    for header_substr in ("id", "mio_model", "path"):
+        assert header_substr in result.output
+
+    if sys.platform == "win32":
+        assert "\u2500" in result.output
+    else:
+        assert "━━" in result.output
+
+    # configs from the temporarily configured test config directory should be included
+    assert "test-wireless-200px" in result.output
+
+    # and configs provided by mio
+    assert "wirefree-sd-layout" in result.output
+
+    # by default paths and mio models should be truncated
+    if sys.platform == "win32":
+        assert "\u2502 .sdcard.SDLayout" in result.output
+        assert "\u2502 wirefree" in result.output
+    else:
+        assert "│ .sdcard.SDLayout" in result.output
+        assert "│ wirefree/" in result.output
+
+    # verbose should display the full values (though truncated in testing because console width is 80)
+    result = runner.invoke(_list, ["-v"], color=False)
+    assert "mio.models." in result.output
+    assert str(CONFIG_DIR)[0:5] in result.output
 
 
 @pytest.mark.timeout(30)